api/pyproject.toml
[tool.poetry]
name = "bit2_back"
version = "0.1.0"
description = "A backend service using hexagonal architecture with FastAPI, web scraping, and LLM integration for Bit2"
authors = ["Ahonakpon Guy Gbaguidi <guygbaguidi123root@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ggbaguidi/bit2_back"
packages = [{include = "bit2_api"}]

[tool.poetry.dependencies]
python = "^3.10"
inject = "^4.3.1"

# API Framework
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
gunicorn = "^21.2.0"
fastapi-versioning = "^0.10.0"
python-multipart = "^0.0.9"
FastAPI-SQLAlchemy = "^0.2.1"

# Database
alembic = "^1.13.1"
psycopg2 = "^2.9.9"


beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
pydantic = "^1.10.2"
pytest-cov = "^6.0.0"
apscheduler = "^3.11.0"
httpx = "^0.28.1"
selenium = "^4.30.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
black = "^22.1.0"
isort = "^5.12.0"
pylint = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py310']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''